% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_curve.R
\name{add_curve}
\alias{add_curve}
\title{Add a fitted linear model to a plot}
\usage{
add_curve(mod, col = "black")
}
\arguments{
\item{mod.}{A linear model with a single response and a single explanatory variable.}

\item{col.}{A character color to use when plotting the line/curve.}
}
\description{
Include a line or curve created from a linear model with a single response
and a single explanatory variable.
}
\examples{
m_line <- lm(weight ~ height, data = cdc)

m_curve
}
\seealso{
\code{\link{add_line}}

\code{\link{lm}}

\code{\link{poly}}
}

