% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargraph.R
\name{bargraph}
\alias{bargraph}
\title{Bargraph Plots}
\usage{
bargraph(x, data, horizontal = FALSE, type = "frequency", ...)
}
\arguments{
\item{x}{formula. An expression of the form \code{~ x | z} where
\code{x} is the variable we want to make bargraphs for and \code{z} is a
second, optional, variable we would like to use for faceting.}

\item{data}{data frame. The data where the variables can be found.}

\item{horizontal}{logical. Use \code{horizontal = TRUE} to make a horizontal
bargraph.}

\item{type}{character. Choose the type of y-axis. Options include \code{"count"}, \code{"frequency"}, \code{"proportion"} and \code{"percent"}.}

\item{groups}{factor. An additional variable to use for side-by-side
bargraphs.}
}
\description{
Create a bargraph of a categorical variable.
}
\section{Note}{
 For additional options for creating bargraphs, look at the
  help file found in the \code{mosaic} package or run \code{help(bargraph,
  package="mosaic")} in the console.
}
\examples{
data(cdc)
bargraph(~depressed, data = cdc)
bargraph(~depressed, data = cdc, groups = gender)
bargraph(~depressed | grade, data = cdc, groups = gender)
}

