% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.R
\name{do}
\alias{do}
\title{A wrapper for the mosaic package's \code{do} function.}
\usage{
do(n)
}
\arguments{
\item{n}{number of times to repeat}
}
\value{
A data frame of the values for the repeated function.
}
\description{
Number of iterations is capped at 500 to prevent individual users from
hogging server computation time. Additional details about the \code{do}
function can be found in the \code{\link[mosaic]{do}} package documentation
}
\section{Note}{
 Find more examples, help and details at the help page for
  \code{\link[mosaic]{do}} in the \code{mosaic} package.
}
\examples{
do(10) * sample(1:10, size=2)
}

