% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{futbol}
\alias{futbol}
\title{Futbol Data Set}
\format{A data frame with _ observations of _ variables}
\source{
\url{}
}
\usage{
data(futbol)
}
\description{
Futbol
}
\details{
\itemize{
 \item
}
}
\keyword{dataset}

