% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/view_data_functions.R
\name{head}
\alias{head}
\alias{tail}
\title{View first/last observations.}
\usage{
head(x, ..., data, groups = NULL, ..fun.. = utils::head,
  na.rm = getOption("na.rm", FALSE))

tail(x, ..., data, groups = NULL, ..fun.. = utils::tail,
  na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{Formula or other object to print.}

\item{data}{Data frame. The data where the variables can be found.}
}
\description{
A pair of functions, altered to understand formula-style syntax. \code{head} will print the first six values of an object, \code{tail} will print the last six.
}
\examples{
head(~hp, data = mtcars)
tail(~hp, data = mtcars)
}

