% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iqr.R
\name{iqr}
\alias{IQR}
\alias{iqr}
\title{IQR computation.}
\usage{
iqr(...)

IQR(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Calculate the IQR of a set of values. The function sets \code{na.rm = TRUE} by default and prints a warning informing users how many values were \code{NA}.
}
\examples{
iqr(~hp, data = mtcars)
IQR(~hp, data = mtcars)
}

