% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_config}
\alias{log_config}
\title{A set of logging functions that Steve made.
These aren't used by the user.}
\usage{
log_config
}
\description{
A set of logging functions that Steve made.
These aren't used by the user.
}
