% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/max.R
\name{max}
\alias{max}
\title{Maximum value computation}
\usage{
max(x, data, ...)
}
\arguments{
\item{x}{an object, often a formula}

\item{data}{a data frame in which to evaluate formulas (or bare names).
Note that the default is \code{data=parent.frame()}.  This makes it convenient to
use this function interactively by treating the working envionment as if it were
a data frame.  But this may not be appropriate for programming uses.
When programming, it is best to use an explicit \code{data} argument
-- ideally supplying a data frame that contains the variables mentioned.}

\item{...}{additional arguments}
}
\description{
Compute the maxiumum of a set of values. The function sets \code{na.rm = TRUE} by default and prints a warning informing users how many values were \code{NA}.
}
\examples{
max(~hp, data = mtcars)
}

