% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{package}
\name{mobilizr}
\alias{mobilizr}
\alias{mobilizr-package}
\title{mobilizr: Functions for the Mobilize Intro to Data Science course.}
\description{
mobilizr provides wrappers and other functions to help students focus on the essentials of learning to code and conduct data science.
}
\details{
To learn more about Mobilize and the Introduction to Data Science course see \href{http://wiki.mobilizingcs.org}{the Mobilize wiki}.
}

