% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{movie}
\alias{movie}
\title{Movie Data Set}
\format{A data frame with 5,043 observations of 16 variables}
\source{
\url{http://www.imdb.com}
}
\usage{
data(movie)
}
\description{
A collection of movies from 1916 to 2016.
}
\details{
\itemize{
\item title. The title of the film.
\item year. The year the film was released.
\item runtime. The length of the film, in minutes.
\item mpaa_rating. The rating of the film as given by the MPAA.
\item color. Is the film in color or black and white.
\item director. Who directed the film
\item aspect_ratio. the ratio of width and height of the film
\item language. The language of the dialog in the film.
\item country. The country the film was produced in.
\item n_poster_faces. The number of faces shown in the movie posters used for advertisment.
\item n_critics. The number of critics who rated the film.
\item n_audeience. The number of regular audience members who rated the film.
\item imdb_score. The rating of the film on IMDB.
\item imdb_id. The unique id of the film on imdb.com.
\item budget. The budget needed to create the film.
\item gross. The amount of money raised by the film.
 }
}
\keyword{dataset}
