% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/view_data_functions.R
\name{print}
\alias{head}
\alias{print}
\alias{tail}
\title{Data viewing functions}
\usage{
print(x, ..., data, groups = NULL, ..fun.. = base::print,
  na.rm = getOption("na.rm", FALSE))

head(x, ..., data, groups = NULL, ..fun.. = utils::head,
  na.rm = getOption("na.rm", FALSE))

tail(x, ..., data, groups = NULL, ..fun.. = utils::tail,
  na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{Formula or other object to print.}

\item{data}{Data frame. The data where the variables can be found.}
}
\description{
A set of functions to view data, altered to understand formula-style syntax. \code{head} will print the first six values of an object, \code{tail} will print the last six and \code{print} will output all values into the console up to a cut-off point.
}
\examples{
print(~hp, data = mtcars)
head(~hp, data = mtcars)
tail(~hp, data = mtcars)
}

