% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHTMLTable.R
\name{readHTMLTable}
\alias{readHTMLTable}
\title{Read data from one or more HTML tables}
\usage{
readHTMLTable(url, which, ...)
}
\arguments{
\item{which}{An integer to choose which table to return. If not specified, return a list of tables.}

\item{...}{see \code{rvest::html_table()}}
}
\description{
Read data from one or more HTML tables
}
\examples{
# Get a list of Tables
url <- "https://www.inflationdata.com/Inflation/Consumer_Price_Index/HistoricalCPI.aspx"
readHTMLTable(url)

# Read the 1st Table
url <- "https://www.inflationdata.com/Inflation/Consumer_Price_Index/HistoricalCPI.aspx"
readHTMLTable(url, which = 1)

}
