% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic passenger data}
\format{A data frame with 1,000 observations of 7 variables}
\usage{
data(titanic)
}
\description{
A dataset containing information on 1,000 randomly sampled passengers of the Titanic.
}
\details{
\itemize{ \item name. name of passenger \item age. age of passenger \item gender. gender of passenger \item fare. amount paid for ticket to sail on the Titanic \item class. the class hospitality/room aboard the Titanic \item embarked. location passenger first stepped on board the Titanic \item survived. the survival status of the passenger }
}
\seealso{
titanic_test
}
\keyword{dataset}

