% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{titanic_test}
\alias{titanic_test}
\title{Titanic passenger survivial test data}
\format{A data frame with 308 observations of 7 variables}
\usage{
data(titanic_test)
}
\description{
A dataset containing information on 308 passengers of the Titanic that are not included in the \code{titanic} data set.
}
\details{
\itemize{ \item name. name of passenger \item age. age of passenger \item gender. gender of passenger \item fare. amount paid for ticket to sail on the Titanic \item class. the class hospitality/room aboard the Titanic \item embarked. location passenger first stepped on board the Titanic \item survived. the survival status of the passenger }
}
\seealso{
titanic
}
\keyword{dataset}

