% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bargraph.R
\name{bargraph}
\alias{bargraph}
\title{Bargraph Plots}
\usage{
bargraph(x, data = parent.frame(), groups, horizontal = FALSE, origin = 0,
  ylab = ifelse(horizontal, "", "Frequency"), xlab = ifelse(horizontal,
  "Frequency", ""), subset, ...)
}
\arguments{
\item{x}{\code{formula}. An expression of the form \code{~ x | z} where \code{x} is the variable we want to make bargraphs for and \code{z} is a second, optional, variable we would like to use for faceting.}

\item{data}{data frame. The data where the variables can be found.}

\item{groups}{factor. An additional variable to use for side-by-side bargraphs.}

\item{horizontal}{logical. Use \code{horizontal = TRUE} to make a horizontal bargraph.}
}
\description{
Create a bargraph.
}
\section{Note}{

For additional options for creating bargraphs, look at the help file found in the \code{mosaic} package or run \code{help(bargraph, package="mosaic")} in the console.
}
\examples{
data(cdc)
bargraph(~depressed, data = cdc)
bargraph(~depressed, data = cdc, groups = gender)
bargraph(~depressed | grade, data = cdc, groups = gender)
}

