% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{cdc}
\alias{cdc}
\title{CDC Youth Risk Behaviour survey}
\format{A data frame with 14,041 observations of 26 variables}
\source{
\url{http://www.cdc.gov/HealthyYouth/yrbs/index.htm}
}
\usage{
data(cdc)
}
\description{
A dataset containing responses from the CDC Youth Risk Behavior Survey. The
variables are:
}
\details{
\itemize{ \item age. age in years \item grade. grade in school \item gender.
sex assigned at birth \item hisp_latino. whether or not student identifies as
Hispanic or Latino \item race. race the student identifies with (A: American
Indian or Alaska Native, B: Asian, C: Black or African American, D: Native
Hawaiian or Other Pacific Islander, E: White, F: Hispanic or Latino, G:
Multiple - Hispanic or Latino, H: Multiple - Non-Hispanic/Latino) \item
height. height of student in meters \item weight. weight of student in
kilograms \item general_health. student perception of their health \item
hours_sleep. reported hours of sleep on school nights \item asthma. student
diagnosed as having asthma \item depressed. student reported feeling
depressed for 2 weeks in a row, or more, during the past 12 months \item
helmet. how often student wore a helmet while bicycling during the past 12
months \item seat_belt. how often student wore a seatbelt in motor vehicle
driven by someone else \item sunscreen. how often student reported wearing
sunscreen on sunny days \item describe_weight. student perception of their
weight relative to what they believe it should be \item eat_fruit. how often
student ate fruit over the previous 7 days \item eat_salad. how often student
ate salad over the previous 7 days \item drink_soda. how often student
consumed soda over the previous 7 days \item drink_milk. how often student
drank milk over the previous 7 days \item days_exercise_20. how often student
was active for at least 20 mins over the previous 7 days. \item
days_exercise_60. how often student was active for at least 60 mins over the
previous 7 days. \item number_teams. number of sports teams played on during
previous 12 months \item hours_tv. average number of hours spent watching tv
on a school day \item hours_videogame. average number of hours spent playing
videogames on a school day \item days_smoking. number of days student
reported smoking cigarettes during past 30 days \item fights. number of
fights student has been during the past 12 months }
}
\keyword{dataset}

