% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{futbol}
\alias{futbol}
\title{Futbol Data Set}
\format{A data frame with 45 observations of 6 variables}
\source{
USMNT data from: \url{http://www.ussoccer.com/mens-national-team/latest-roster} NFL data from: \url{http://www.besttickets.com/blog/wp-content/uploads/2014/09/2014-NFL-Census-Shared.xls}
}
\usage{
data(futbol)
}
\description{
Player data for National Football League players from 2014 or from members of
the U.S. Men's National Soccer Team from 2014.
}
\details{
\itemize{ \item team. The home city for each player's team. \item player. The
name of the player. \item ht_inches. The player's height in inches. \item
wt_lbs. The player's weight in pounds. \item age. The age of the player in
2014. \item league. Indicates whether the player plays for the U.S. Mens
National Soccer Team or the NFL.}
}
\keyword{dataset}
