% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{package}
\name{mobilizr}
\alias{mobilizr}
\title{mobilizr: Functions for the Mobilize Intro to Data Science course.}
\description{
mobilizr provides wrappers and other functions to help students focus on the
essentials of learning to code and conduct data science.
}
\details{
To learn more about Mobilize and the Introduction to Data Science course see
\href{http://www.mobilizingcs.org}{the Mobilize homepage}.
}
