% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{movie}
\alias{movie}
\title{Movie Data Set}
\format{A data frame with _ observations of _ variables}
\source{
\url{http://www.rottentomatos.com}
}
\usage{
data(movie)
}
\description{
Movie
}
\details{
\itemize{
 \item
}
}
\keyword{dataset}

